Cleaning up some code and providing more interfaces for the UI.

+++ trunk/src/utilities/Parameters.java
@@ -19,8 +19,8 @@ public class Parameters {
 	public int mHeight;
 	
 	// Dimensions of each segment
-	public int sWidth;
-	public int sHeight;
+	public float sWidth;
+	public float sHeight;
 	
 	private boolean initialized;
